fix sign compare errors on OS X builds

+++ include/microscopes/common/special.hpp
@@ -16,7 +16,7 @@ struct special {
     MICROSCOPES_ASSERT(d > 0);
     const float term1 = 0.25*float(d*(d-1))*1.1447298858494002 /* log(pi) */;
     float term2 = 0.;
-    for (int j = 1; j <= d; j++)
+    for (int j = 1; j <= (int)d; j++)
       term2 += distributions::fast_lgamma(a + 0.5*float(1-j));
     return term1 + term2;
   }
